requirements.txt
anyio==3.7.1
boto3==1.28.25
botocore==1.31.25
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
environs==9.5.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
jira==3.5.2
jmespath==1.0.1
marshmallow==3.20.1
notion-client==2.0.0
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==1.26.16
